refactor(admin): isolate cross-workspace task reads under /admin
Move the admin's ability to view any task off the user-facing /tasks/[taskId]
route and into /admin/tasks/[taskId], so elevated cross-workspace reads live
only behind the admin layout's requireAdminSession guard.

- extract the task render tree into a shared read-only-capable TaskDetailView;
  add forceReadOnly (admin locks all mutations) and enableAutoSwitch (user-only
  workspace follow)
- /tasks/[taskId] is a thin wrapper over the workspace-scoped read again; the
  admin fallback and isAdminSession helper are removed, so /tasks/* can never
  resolve a foreign task
- /admin/tasks/[taskId] renders the full read-only view via the admin endpoint
  plus an admin context strip (owner email, organization, back); the redundant
  Open button and its unused i18n keys are dropped
