+++ apps/web/messages/fr.json
@@ -2702,13 +2702,7 @@
         },
         "TaskDetail": {
           "backToList": "Retour aux tâches",
-          "created": "Créée",
-          "open": "Ouvrir",
-          "organization": "Organisation",
-          "personalWorkspace": "Espace de travail personnel",
-          "status": "Statut",
-          "taskId": "ID de la tâche",
-          "user": "Utilisateur"
+          "personalWorkspace": "Espace de travail personnel"
         }
       }
     },
