+++ apps/web/messages/pt.json
@@ -2702,13 +2702,7 @@
         },
         "TaskDetail": {
           "backToList": "Voltar às tarefas",
-          "created": "Criada",
-          "open": "Abrir",
-          "organization": "Organização",
-          "personalWorkspace": "Espaço de trabalho pessoal",
-          "status": "Estado",
-          "taskId": "ID da tarefa",
-          "user": "Utilizador"
+          "personalWorkspace": "Espaço de trabalho pessoal"
         }
       }
     },
