+++ apps/web/messages/zh-Hans.json
@@ -2702,13 +2702,7 @@
         },
         "TaskDetail": {
           "backToList": "返回任务列表",
-          "created": "创建时间",
-          "open": "打开",
-          "organization": "组织",
-          "personalWorkspace": "个人工作区",
-          "status": "状态",
-          "taskId": "任务 ID",
-          "user": "用户"
+          "personalWorkspace": "个人工作区"
         }
       }
     },
