+++ apps/web/src/app/(app)/tasks/components/task-detail-view.tsx
@@ -1,0 +1,551 @@
+import type { SubscriptionPlanName } from "@sokosumi/utils";
+import { resolveIpfsOrHttpUrl } from "@sokosumi/utils";
+import { getLocale, getTranslations } from "next-intl/server";
+import { Suspense } from "react";
+import { AutoContextSwitch } from "@/app/components/auto-context-switch";
+import { TaskActivitySection } from "@/app/tasks/components/task-activity";
+import { TaskDescription } from "@/app/tasks/components/task-description";
+import { TaskDetailActions } from "@/app/tasks/components/task-detail-actions";
+import { mapVisibleTaskLinks } from "@/app/tasks/components/task-detail-api-types";
+import { TaskDetailHeader } from "@/app/tasks/components/task-detail-header";
+import { TaskJobs } from "@/app/tasks/components/task-jobs";
+import { TaskMetadata } from "@/app/tasks/components/task-metadata";
+import { TaskRelatedTasks } from "@/app/tasks/components/task-related-tasks";
+import { TaskStatusRealtimeListener } from "@/app/tasks/components/task-status-realtime-listener";
+import { buildAgentNameById } from "@/app/tasks/utils/agent-names";
+import { getCoworkerOptions } from "@/app/tasks/utils/coworker-options";
+import { buildTaskActivityActors } from "@/app/tasks/utils/task-activity-actors";
+import { buildTaskStatusLabels } from "@/app/tasks/utils/task-status-labels";
+import { parsePlanName } from "@/components/billing/subscription-plan-utils";
+import { getSession } from "@/lib/auth/utils";
+import { CoreApiRequestError, coreClient } from "@/lib/clients/core.client";
+import type { Task } from "@/lib/clients/generated/core/types.gen";
+import { agentService } from "@/lib/services";
+import { coworkerService } from "@/lib/services/coworker.service";
+import { projectService } from "@/lib/services/project.service";
+import { userService } from "@/lib/services/user.service";
+import { resolveAccountName } from "@/lib/utils/account-name";
+import { mapTaskToTaskWithCoworker } from "@/lib/utils/task-transformer";
+
+type SessionResult = Awaited<ReturnType<typeof getSession>>;
+type AgentsResult = Awaited<
+  ReturnType<typeof agentService.getAvailableAgentsWithCreditsPrice>
+>;
+type CoworkersResult = Awaited<
+  ReturnType<typeof coworkerService.listCoworkers>
+>;
+type MembersResult = Awaited<
+  ReturnType<typeof userService.getMyMembersWithOrganizations>
+>;
+
+interface TaskDetailViewProps {
+  task: Task;
+  /**
+   * Forces the view into read-only mode regardless of ownership. Used by the
+   * admin task detail page, where the viewer is never the owner and must not be
+   * able to edit, comment, or mutate the task.
+   */
+  forceReadOnly?: boolean;
+  /**
+   * Renders {@link AutoContextSwitch} so the active workspace follows the task.
+   * Only safe when the viewer is guaranteed to be a member of the task's
+   * workspace (the user-facing route). The admin route leaves this off — admins
+   * read tasks in workspaces they are not part of, where switching would fail or
+   * be wrong.
+   */
+  enableAutoSwitch?: boolean;
+}
+
+/**
+ * Read-only-capable task detail view shared by the user-facing
+ * `/tasks/[taskId]` route and the admin `/admin/tasks/[taskId]` route. The user
+ * route loads the task through the workspace-scoped Core read; the admin route
+ * loads it through the admin-only endpoint and passes `forceReadOnly`.
+ */
+export async function TaskDetailView({
+  task,
+  forceReadOnly = false,
+  enableAutoSwitch = false,
+}: TaskDetailViewProps) {
+  const taskId = task.id;
+  const coworkersPromise = coworkerService.listCoworkers();
+  const agentsPromise = agentService.getAvailableAgentsWithCreditsPrice();
+  const membersPromise = userService.getMyMembersWithOrganizations();
+  const sessionPromise = getSession();
+  const localePromise = getLocale();
+  const currentPlanPromise = sessionPromise.then((session) =>
+    getCurrentPlan(session, task.organizationId),
+  );
+  const translationsPromise = getTranslations("App.Tasks.Detail");
+  const linkedTasks = mapVisibleTaskLinks(task.links);
+
+  const t = await translationsPromise;
+
+  return (
+    <div className="min-h-full w-full">
+      <div className="mx-auto max-w-4xl px-4 pb-8">
+        {enableAutoSwitch ? (
+          <Suspense fallback={null}>
+            <TaskDetailEffects
+              taskId={taskId}
+              targetOrganizationId={task.workspace.organizationId ?? null}
+              membersPromise={membersPromise}
+              sessionPromise={sessionPromise}
+            />
+          </Suspense>
+        ) : null}
+
+        <TaskDetailHeader
+          taskName={task.name}
+          backLabel={t("back")}
+          actions={
+            <Suspense fallback={<TaskDetailActionsFallback />}>
+              <TaskDetailActionsSlot
+                taskId={taskId}
+                task={task}
+                forceReadOnly={forceReadOnly}
+                coworkersPromise={coworkersPromise}
+                agentsPromise={agentsPromise}
+                membersPromise={membersPromise}
+                sessionPromise={sessionPromise}
+              />
+            </Suspense>
+          }
+        />
+
+        <div className="mt-6 space-y-8">
+          <Suspense
+            fallback={
+              <TaskOverviewFallback
+                descriptionTitle={t("description")}
+                propertiesTitle={t("properties")}
+              />
+            }
+          >
+            <TaskOverviewSection
+              task={task}
+              coworkersPromise={coworkersPromise}
+              agentsPromise={agentsPromise}
+            />
+          </Suspense>
+
+          <TaskRelatedTasks
+            title={t("linkedTasksTitle")}
+            emptyLabel={t("linkedTasksEmpty")}
+            tasks={linkedTasks}
+            relationLabels={{
+              related: t("actions.relations.related"),
+              blocks: t("actions.relations.blocks"),
+              blocked_by: t("actions.relations.blockedBy"),
+              parent: t("actions.relations.subtask"),
+              child: t("actions.relations.parent"),
+              duplicate: t("actions.relations.duplicate"),
+            }}
+          />
+
+          {task.jobs.length > 0 && (
+            <>
+              <Suspense
+                fallback={<TaskSectionFallback title={t("jobs")} rows={3} />}
+              >
+                <TaskJobsSection
+                  task={task}
+                  agentsPromise={agentsPromise}
+                  sessionPromise={sessionPromise}
+                  localePromise={localePromise}
+                />
+              </Suspense>
+            </>
+          )}
+          <Suspense
+            fallback={<TaskSectionFallback title={t("activity")} rows={4} />}
+          >
+            <TaskActivitySectionContent
+              taskId={taskId}
+              task={task}
+              forceReadOnly={forceReadOnly}
+              agentsPromise={agentsPromise}
+              sessionPromise={sessionPromise}
+              currentPlanPromise={currentPlanPromise}
+            />
+          </Suspense>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+/**
+ * Read-only unless the viewer owns the task. Workspace collaborators (org task,
+ * not the owner) are read-only, matching the original user-route behavior;
+ * `forceReadOnly` additionally locks the view for admins on any task.
+ */
+function isReadOnlyForViewer(
+  task: Task,
+  session: SessionResult,
+  forceReadOnly: boolean,
+): boolean {
+  if (forceReadOnly) {
+    return true;
+  }
+  return (
+    task.workspace.organizationId !== null && session?.user.id !== task.userId
+  );
+}
+
+async function TaskDetailEffects({
+  taskId,
+  targetOrganizationId,
+  membersPromise,
+  sessionPromise,
+}: {
+  taskId: string;
+  targetOrganizationId: string | null;
+  membersPromise: Promise<MembersResult>;
+  sessionPromise: Promise<SessionResult>;
+}) {
+  const [members, session, t, tOrganizationSwitcher] = await Promise.all([
+    membersPromise,
+    sessionPromise,
+    getTranslations("App.Tasks.Detail"),
+    getTranslations("Components.OrganizationSwitcher"),
+  ]);
+  const activeOrganizationId = session?.session.activeOrganizationId ?? null;
+  const targetAccountName = resolveAccountName(
+    targetOrganizationId,
+    members,
+    tOrganizationSwitcher("personalAccount"),
+  );
+
+  return (
+    <>
+      <AutoContextSwitch
+        activeOrganizationId={activeOrganizationId}
+        targetOrganizationId={targetOrganizationId}
+        successMessage={t("switchedWorkspace", {
+          account: targetAccountName,
+        })}
+      />
+      {session?.user.id ? (
+        <TaskStatusRealtimeListener userId={session.user.id} taskId={taskId} />
+      ) : null}
+    </>
+  );
+}
+
+async function TaskOverviewSection({
+  task,
+  coworkersPromise,
+  agentsPromise,
+}: {
+  task: Task;
+  coworkersPromise: Promise<CoworkersResult>;
+  agentsPromise: Promise<AgentsResult>;
+}) {
+  const projectPromise = task.projectId
+    ? projectService.getProjectById(task.projectId).catch(() => null)
+    : Promise.resolve(null);
+  const [coworkers, agents, project, t, tStatus] = await Promise.all([
+    coworkersPromise,
+    agentsPromise,
+    projectPromise,
+    getTranslations("App.Tasks.Detail"),
+    getTranslations("App.Tasks.Filters.statusOptions"),
+  ]);
+  const { task: taskWithCoworker, agentNameById } = buildTaskDetailContext(
+    task,
+    coworkers,
+    agents,
+  );
+
+  return (
+    <>
+      <TaskDescription
+        title={t("description")}
+        description={taskWithCoworker.description}
+        agentNameById={agentNameById}
+        expandLabel={t("expand")}
+        collapseLabel={t("collapse")}
+      />
+
+      <TaskMetadata
+        task={task}
+        project={project ? { id: project.id, name: project.name } : null}
+        labels={{
+          propertiesTitle: t("properties"),
+          status: t("status"),
+          statusLabels: buildTaskStatusLabels((key) => tStatus(key)),
+          owner: t("owner"),
+          organization: t("organization"),
+          personalWorkspace: t("personalWorkspace"),
+          project: t("project"),
+          coworker: t("coworker"),
+          created: t("created"),
+          updated: t("updated"),
+        }}
+      />
+    </>
+  );
+}
+
+async function TaskDetailActionsSlot({
+  taskId,
+  task,
+  forceReadOnly,
+  coworkersPromise,
+  agentsPromise,
+  membersPromise,
+  sessionPromise,
+}: {
+  taskId: string;
+  task: Task;
+  forceReadOnly: boolean;
+  coworkersPromise: Promise<CoworkersResult>;
+  agentsPromise: Promise<AgentsResult>;
+  membersPromise: Promise<MembersResult>;
+  sessionPromise: Promise<SessionResult>;
+}) {
+  const [coworkers, agents, members, session, t, tMembersTableHeader] =
+    await Promise.all([
+      coworkersPromise,
+      agentsPromise,
+      membersPromise,
+      sessionPromise,
+      getTranslations("App.Tasks.Detail"),
+      getTranslations("Components.MembersTable.Header"),
+    ]);
+  const {
+    task: taskWithCoworker,
+    agentNameById,
+    coworkerOptions,
+  } = buildTaskDetailContext(task, coworkers, agents);
+  const isReadOnlyWorkspaceView = isReadOnlyForViewer(
+    task,
+    session,
+    forceReadOnly,
+  );
+  const personalWorkspaceMoveLabel =
+    session?.user?.name?.trim() ||
+    session?.user?.email?.trim() ||
+    t("actions.personalWorkspace");
+
+  return (
+    <TaskDetailActions
+      share={taskWithCoworker.share ?? null}
+      taskId={taskId}
+      status={taskWithCoworker.status}
+      jobsCount={taskWithCoworker.jobsCount}
+      taskLinks={task.links}
+      coworkerOptions={coworkerOptions}
+      agentNameById={agentNameById}
+      defaultCoworkerId={task.coworkerId}
+      currentOrganizationId={task.workspace.organizationId ?? null}
+      organizations={members}
+      personalWorkspaceLabel={personalWorkspaceMoveLabel}
+      isReadOnly={isReadOnlyWorkspaceView}
+      actionsMenuLabel={tMembersTableHeader("actions")}
+      labels={{
+        edit: t("actions.edit"),
+        archive: t("actions.archive"),
+        confirmArchive: t("actions.confirmArchive"),
+        confirmArchiveDescription: t("actions.confirmArchiveDescription"),
+        archiveError: t("actions.archiveError"),
+        markAsReady: t("actions.markAsReady"),
+        revertToDraft: t("actions.revertToDraft"),
+        cancelRequest: t("actions.cancelRequest"),
+        share: t("actions.share"),
+      }}
+    />
+  );
+}
+
+async function TaskJobsSection({
+  task,
+  agentsPromise,
+  sessionPromise,
+  localePromise,
+}: {
+  task: Task;
+  agentsPromise: Promise<AgentsResult>;
+  sessionPromise: Promise<SessionResult>;
+  localePromise: Promise<string>;
+}) {
+  const [agents, session, locale, t] = await Promise.all([
+    agentsPromise,
+    sessionPromise,
+    localePromise,
+    getTranslations("App.Tasks.Detail"),
+  ]);
+
+  return (
+    <TaskJobs
+      title={t("jobs")}
+      agents={agents}
+      jobs={task.jobs}
+      userId={session?.user.id ?? null}
+      locale={locale}
+      emptyLabel={t("jobsEmpty")}
+      untitledLabel={t("jobsUntitled")}
+      unknownAgentLabel={t("jobsUnknownAgent")}
+    />
+  );
+}
+
+async function TaskActivitySectionContent({
+  taskId,
+  task,
+  forceReadOnly,
+  agentsPromise,
+  sessionPromise,
+  currentPlanPromise,
+}: {
+  taskId: string;
+  task: Task;
+  forceReadOnly: boolean;
+  agentsPromise: Promise<AgentsResult>;
+  sessionPromise: Promise<SessionResult>;
+  currentPlanPromise: Promise<SubscriptionPlanName>;
+}) {
+  const [agents, session, currentPlan, t] = await Promise.all([
+    agentsPromise,
+    sessionPromise,
+    currentPlanPromise,
+    getTranslations("App.Tasks.Detail"),
+  ]);
+  const { userById: actorsUserById, coworkerById } =
+    buildTaskActivityActors(task);
+  const currentUser = session?.user
+    ? {
+        id: session.user.id,
+        name: session.user.name ?? "User",
+        image: session.user.image
+          ? resolveIpfsOrHttpUrl(session.user.image)
+          : null,
+      }
+    : null;
+  const userById = currentUser
+    ? {
+        ...actorsUserById,
+        [currentUser.id]: {
+          name: currentUser.name,
+          image: currentUser.image,
+        },
+      }
+    : actorsUserById;
+  const agentNameById = buildAgentNameById(agents);
+  const isFreePlan = currentPlan === "free";
+  const isReadOnlyWorkspaceView = isReadOnlyForViewer(
+    task,
+    session,
+    forceReadOnly,
+  );
+
+  return (
+    <TaskActivitySection
+      taskId={taskId}
+      title={t("activity")}
+      placeholder={t("commentPlaceholder")}
+      attachLabel={t("attach")}
+      submitLabel={t("submit")}
+      actorCoworkerLabel={t("actorCoworker")}
+      actorUserLabel={t("actorUser")}
+      actorSystemLabel={t("actorSystem")}
+      actionCommentedLabel={t("actionCommented")}
+      actionUpdatedStatusLabel={t("actionUpdatedStatus")}
+      events={task.events}
+      agentNameById={agentNameById}
+      userById={userById}
+      coworkerById={coworkerById}
+      currentUser={currentUser}
+      expandLabel={t("expand")}
+      collapseLabel={t("collapse")}
+      isFreePlan={isFreePlan}
+      canComment={!isReadOnlyWorkspaceView}
+    />
+  );
+}
+
+async function getCurrentPlan(
+  session: SessionResult,
+  organizationId: string | null,
+): Promise<SubscriptionPlanName> {
+  if (!session) {
+    return "free";
+  }
+
+  try {
+    const { data } = organizationId
+      ? await coreClient.getOrganizationActiveSubscription(organizationId)
+      : await coreClient.getMyActiveSubscription();
+
+    return parsePlanName(data.subscription?.plan) ?? "free";
+  } catch (error) {
+    // Tasks can be viewed in workspaces the caller can no longer resolve a
+    // subscription for (e.g. revoked membership) — treat those as free.
+    if (
+      error instanceof CoreApiRequestError &&
+      (error.status === 403 || error.status === 404)
+    ) {
+      return "free";
+    }
+    throw error;
+  }
+}
+
+function buildTaskDetailContext(
+  task: Task,
+  coworkers: CoworkersResult,
+  agents: AgentsResult,
+) {
+  const coworkersById = new Map(
+    coworkers.map((coworker) => [coworker.id, coworker]),
+  );
+  const agentsById = new Map(agents.map((agent) => [agent.id, agent]));
+
+  return {
+    task: mapTaskToTaskWithCoworker(task, coworkersById, agentsById),
+    agentNameById: buildAgentNameById(agents),
+    coworkerOptions: getCoworkerOptions(coworkers),
+  };
+}
+
+function TaskDetailActionsFallback() {
+  return <div className="bg-muted h-9 w-9 animate-pulse rounded-md" />;
+}
+
+function TaskOverviewFallback({
+  descriptionTitle,
+  propertiesTitle,
+}: {
+  descriptionTitle: string;
+  propertiesTitle: string;
+}) {
+  return (
+    <>
+      <TaskSectionFallback title={descriptionTitle} rows={3} />
+      <TaskSectionFallback title={propertiesTitle} rows={4} />
+    </>
+  );
+}
+
+function TaskSectionFallback({
+  title,
+  rows = 3,
+}: {
+  title: string;
+  rows?: number;
+}) {
+  return (
+    <section className="space-y-4">
+      <h2 className="text-muted-foreground/60 text-xs font-medium">{title}</h2>
+      <div className="space-y-3">
+        {Array.from({ length: rows }, (_, index) => (
+          <div
+            key={`${title}-${index}`}
+            className="bg-muted h-4 animate-pulse rounded"
+          />
+        ))}
+      </div>
+    </section>
+  );
+}
